Remove old code
Fixed: The Windows refresh script, that refreshes the Bibledit core
library, does not now check that *.h and *.cpp files no longer in the
core but still in the Windows port, will be removed from that Windows
port. As a result, there's old code in the Windows port that's no
longer in the core library. This fixes
https://github.com/bibledit/bibledit/issues/524