Fix for too small a window
This fixes https://github.com/bibledit/cloud/issues/43: Windows client
window too small. On Windows (10) it was observed that somehow the
screen got too small to be visible. Only "Maximizing" the window brought
it back to visibility. Needs to be looked into why it can become too
small. Perhaps the settings for the screen size on Windows need to be
limited, so it does not get too small.